// src/models/Profile2.ts
// export class Profile2 {
//     id?: number;
//     username: string = "";
//     first_name: string = "";
//     last_name: string = "";
//     password: string = "";
//     email: string = "";
//     items222: number[] = [];
//     is_staff: boolean = false;
//   }

export interface Profile2 {
    id?: number;
    first_name: string;
    last_name: string;
    username: string;
    password: string;
    email: string;
    is_staff: boolean;
    items222: Number[];
  }